Extracted crc to own test suite
